ФЗ 11. FilmorateApplication. Добавлена возможность лайкать фильмы и добавляться в друзья. Выводить топ-10 фильмов. Исправлены замечания
