+++ src/main/java/ru/yandex/practicum/filmorate/controller/FilmController.java
@@ -1,8 +1,8 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import ru.yandex.practicum.filmorate.model.Film;
@@ -13,15 +13,9 @@ import java.util.Collection;
 @RestController
 @RequestMapping("/films")
 @Slf4j
+@RequiredArgsConstructor
 public class FilmController {
     private final FilmService filmService;
-
-    @Autowired
-    public FilmController(FilmService filmService) {
-        log.debug("FilmController({}).", filmService.getClass().getSimpleName());
-        this.filmService = filmService;
-        log.info("Подключена зависимость: {}.", filmService.getClass().getName());
-    }
 
     @GetMapping
     public Collection<Film> getAllFilms() {
