+++ src/main/java/ru/yandex/practicum/filmorate/controller/UserController.java
@@ -1,8 +1,8 @@
 package ru.yandex.practicum.filmorate.controller;
 
 import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
 import ru.yandex.practicum.filmorate.model.User;
@@ -13,15 +13,9 @@ import java.util.Collection;
 @Slf4j
 @RestController
 @RequestMapping("/users")
+@RequiredArgsConstructor
 public class UserController {
     private final UserService userService;
-
-    @Autowired
-    public UserController(UserService userService) {
-        log.debug("UserController({}).", userService.getClass().getSimpleName());
-        this.userService = userService;
-        log.info("Подключена зависимость: {}.", userService.getClass().getName());
-    }
 
     @GetMapping
     public Collection<User> getAllUsers() {
