+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -1,14 +1,16 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
+import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
+import java.time.LocalDate;
 import java.util.Collection;
 import java.util.Comparator;
 import java.util.List;
@@ -17,26 +19,20 @@ import java.util.stream.Collectors;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class FilmService {
     private final FilmStorage filmStorage;
     private final UserStorage userStorage;
 
-    @Autowired
-    public FilmService(FilmStorage filmStorage, UserStorage userStorage) {
-        log.debug("FilmService({}, {})",
-                filmStorage.getClass().getSimpleName(),
-                userStorage.getClass().getSimpleName());
-        this.filmStorage = filmStorage;
-        log.info("Подключена зависимость: {}.", filmStorage.getClass().getName());
-        this.userStorage = userStorage;
-        log.info("Подключена зависимость: {}.", userStorage.getClass().getName());
-    }
-
     public Film create(Film film) {
+        validateFilm(film);
         return filmStorage.create(film);
     }
 
     public Film update(Film film) {
+        if (film.getId() == 0) {
+            throw new ValidationException("ID не может быть равен 0");
+        }
         return filmStorage.update(film);
     }
 
@@ -57,52 +53,61 @@ public class FilmService {
     }
 
     public Collection<Film> getPopularFilms(int count) {
-        log.debug("getPopularFilms({}).", count);
-         List<Film> popularFilms =
-                filmStorage.getAllValues()
+        List<Film> popularFilms =
+                getAllValues()
                         .stream()
                         .sorted(Comparator.comparingInt(this::getLikeCount).reversed())
                         .limit(count)
                         .collect(Collectors.toList());
-        log.trace("Возвращены популярные фильмы: {}.", popularFilms);
         return popularFilms;
     }
 
     private int getLikeCount(Film film) {
-        log.trace("getLikeCount({}).", film);
-        log.trace("Возвращено кол-во лайков: {}.", film.getLikes().size());
         return film.getLikes().size();
     }
 
     public void addLike(Long filmId, Long userId) {
-        log.debug("addLike({}, {}).", filmId, userId);
-        if (!filmStorage.getAll().containsKey(filmId)) {
+        User user = userStorage.getById(userId);
+        Film film = getById(filmId);
+        if (film == null) {
             throw new NotFoundException("Фильм с ID " + filmId + " не найден.");
         }
-        if (!userStorage.getAll().containsKey(userId)) {
+        if (user == null) {
             throw new NotFoundException("Пользователь с ID " + userId + " не найден.");
         }
-        if (filmStorage.getAll().get(filmId).getLikes().contains(userId)) {
+        if (film.getLikes().contains(userId)) {
             throw new ValidationException("Пользователь с ID " + userId + " уже добавлял like");
         }
-        filmStorage.getAll().get(filmId).getLikes().add(userId);
-        log.debug("Добавлен лайк: {}.", filmStorage.getAll().get(filmId));
+        film.getLikes().add(userId);
     }
 
     public void deleteLike(Long filmId, Long userId) {
-        log.debug("deleteLike({}, {}).", filmId, userId);
-        if (!filmStorage.getAll().containsKey(filmId)) {
+        User user = userStorage.getById(userId);
+        Film film = getById(filmId);
+        if (film == null) {
             throw new NotFoundException("Фильм с ID " + filmId + " не найден.");
         }
-        if (!userStorage.getAll().containsKey(userId)) {
+        if (user == null) {
             throw new NotFoundException("Пользователь с ID " + userId + " не найден.");
         }
-        if (!filmStorage.getAll().get(filmId).getLikes().contains(userId)) {
-            throw new ValidationException("Пользователь с ID " + userId + " не добавлял like");
-        }
-        filmStorage.getAll().get(filmId).getLikes().remove(userId);
-        log.debug("Лайк удалён: {}.", filmStorage.getAll().get(filmId));
+        film.getLikes().remove(userId);
     }
 
-
+    private void validateFilm(Film film) {
+        if (film.getName() == null || film.getName().isBlank()) {
+            throw new ValidationException("Имя не должно быть пустым");
+        }
+        if (film.getDescription() != null && film.getDescription().length() > 200) {
+            throw new ValidationException("Длина описания не должна превышать 200 символов");
+        }
+        if (film.getReleaseDate() == null || film.getReleaseDate().isBefore(LocalDate.of(1895, 12, 28))) {
+            throw new ValidationException("Дата релиза не должна быть ранее 28.12.1895");
+        }
+        if (film.getReleaseDate().isAfter(LocalDate.now())) {
+            throw new ValidationException("Дата релиза не может быть в будущем");
+        }
+        if (film.getDuration() <= 0) {
+            throw new ValidationException("Продолжительность фильма должна быть положительным числом");
+        }
+    }
 }
