+++ src/main/java/ru/yandex/practicum/filmorate/service/UserService.java
@@ -1,33 +1,32 @@
 package ru.yandex.practicum.filmorate.service;
 
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.User;
 import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
+import java.time.LocalDate;
 import java.util.*;
 import java.util.stream.Collectors;
 
 @Slf4j
 @Service
+@RequiredArgsConstructor
 public class UserService {
     private final UserStorage userStorage;
 
-    @Autowired
-    public UserService(UserStorage userStorage) {
-        log.debug("UserService({}).", userStorage.getClass().getSimpleName());
-        this.userStorage = userStorage;
-        log.info("Подключена зависимость: {}.", userStorage.getClass().getName());
-    }
-
     public User create(User user) {
+        validateUser(user);
         return userStorage.create(user);
     }
 
     public User update(User user) {
+        if (user.getId() == 0) {
+            throw new ValidationException("ID не может быть равен 0");
+        }
         return userStorage.update(user);
     }
 
@@ -53,7 +52,6 @@ public class UserService {
         if (user != null && friend != null) {
             user.getFriends().add(friendId);
             friend.getFriends().add(userId);
-            log.info("Пользователи {} и {} теперь друзья", user, friend);
         }
         if (user == null) {
             throw new NotFoundException("Пользователь с ID " + userId + " не найден.");
@@ -79,9 +77,7 @@ public class UserService {
         if (friend == null) {
             throw new NotFoundException("Пользователь с ID " + friendId + " не найден.");
         }
-        /*if (!user.getFriends().contains(friendId) || !friend.getFriends().contains(userId)) {
-            throw new ValidationException("Пользователи не являются друзьями " + user +" " + friend);
-        }*/
+
         user.getFriends().remove(friendId);
         friend.getFriends().remove(userId);
         log.info("Пользователи {} и {} теперь не являются друзьями", user, friend);
@@ -102,5 +98,21 @@ public class UserService {
         return user.getFriends().stream()
                 .map(this::getById)
                 .collect(Collectors.toList());
+    }
+
+    private void validateUser(User user) {
+        if (user.getEmail() == null || user.getEmail().isBlank() || !user.getEmail().contains("@")) {
+            throw new ValidationException("Электронная почта не может быть пустой и должна содержать знак - @");
+        }
+        if (user.getLogin() == null || user.getLogin().isBlank() || user.getLogin().contains(" ")) {
+            throw new ValidationException("Логин не может быть пустым или содержать пробелы");
+        }
+        LocalDate toDay = LocalDate.now();
+        if (user.getBirthday() == null || user.getBirthday().isAfter(toDay)) {
+            throw new ValidationException("Дата рождения не может быть больше текущей даты");
+        }
+        if (user.getName() == null || user.getName().isBlank()) {
+            user.setName(user.getLogin());
+        }
     }
 }
