+++ src/main/java/ru/yandex/practicum/filmorate/storage/film/InMemoryFilmStorage.java
@@ -3,10 +3,8 @@ package ru.yandex.practicum.filmorate.storage.film;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Component;
 import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 
-import java.time.LocalDate;
 import java.util.Collection;
 import java.util.HashMap;
 import java.util.Map;
@@ -19,7 +17,6 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film create(Film film) {
-        validateFilm(film);
         idFilm++;
         film.setId(idFilm);
         films.put(film.getId(), film);
@@ -49,9 +46,6 @@ public class InMemoryFilmStorage implements FilmStorage {
 
     @Override
     public Film update(Film film) {
-        if (film.getId() == 0) {
-            throw new ValidationException("ID не может быть равен 0");
-        }
         if (films.containsKey(film.getId())) {
             films.put(film.getId(), film);
         } else {
@@ -66,24 +60,6 @@ public class InMemoryFilmStorage implements FilmStorage {
             films.remove(id);
         } else {
             throw new NotFoundException("Фильм с ID " + id + " не найден.");
-        }
-    }
-
-    private void validateFilm(Film film) {
-        if (film.getName() == null || film.getName().isBlank()) {
-            throw new ValidationException("Имя не должно быть пустым");
-        }
-        if (film.getDescription() != null && film.getDescription().length() > 200) {
-            throw new ValidationException("Длина описания не должна превышать 200 символов");
-        }
-        if (film.getReleaseDate() == null || film.getReleaseDate().isBefore(LocalDate.of(1895,12,28))) {
-            throw new ValidationException("Дата релиза не должна быть ранее 28.12.1895");
-        }
-        if (film.getReleaseDate().isAfter(LocalDate.now())) {
-            throw new ValidationException("Дата релиза не может быть в будущем");
-        }
-        if (film.getDuration() <= 0) {
-            throw new ValidationException("Продолжительность фильма должна быть положительным числом");
         }
     }
 }
